notification-service/requirements.txt
fastapi
uvicorn[standard]
pika
python-dotenv
requests
